# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MarketMitra API",
    "version": "1.0.0",
    "description": "Public + session-scoped HTTP API for the MarketMitra dashboard. For automated / agent access to public market data, prefer the MCP server at /api/mcp (see /llms.txt and docs/api-surface.md). Session-scoped endpoints authenticate via the Clerk session cookie in hosted mode; in self-host mode they resolve to a single local user with no login."
  },
  "servers": [
    { "url": "https://marketmitra-v2.vercel.app", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local dev" }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__session",
        "description": "Clerk session cookie (hosted mode). Sent automatically by the browser; the API explorer uses your current session."
      },
      "cronBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET bearer token."
      }
    },
...
  },
  "paths": {
...
    "/api/notes": {
      "get": {
        "tags": ["Portfolio"],
        "summary": "List notes",
        "description": "The user's saved notes (Phase 10 / ADR 0020), newest-updated first. Notes also feed the user's private retrieval corpus.",
        "security": [{ "cookieAuth": [] }],
        "responses": { "200": { "description": "The user's notes.", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } }, "401": { "description": "Not authenticated." } }
      },
      "post": {
        "tags": ["Portfolio"],
        "summary": "Create a note",
        "security": [{ "cookieAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["title", "body"], "properties": { "title": { "type": "string", "maxLength": 140 }, "body": { "type": "string", "maxLength": 4000 }, "symbol": { "type": "string", "nullable": true } } }, "example": { "title": "Why I hold TCS", "body": "Steady dividend, strong order book.", "symbol": "TCS" } } } },
        "responses": { "201": { "description": "Created." }, "401": { "description": "Not authenticated." }, "409": { "description": "Per-user note limit (200) reached." }, "422": { "description": "Invalid input." } }
      }
    },
    "/api/notes/{id}": {
      "patch": {
        "tags": ["Portfolio"],
        "summary": "Update a note",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["title", "body"], "properties": { "title": { "type": "string" }, "body": { "type": "string" }, "symbol": { "type": "string", "nullable": true } } } } } },
        "responses": { "200": { "description": "Updated." }, "401": { "description": "Not authenticated." }, "404": { "description": "Not found / not owned." } }
      },
      "delete": {
        "tags": ["Portfolio"],
        "summary": "Delete a note",
        "security": [{ "cookieAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Deleted." }, "401": { "description": "Not authenticated." }, "404": { "description": "Not found / not owned." } }
      }
    },
...
  }
}